[BENCH] updated benchmark with quantization implementation

+++ 00_graph_bench/include/utils/graphTest.h
@@ -10,7 +10,7 @@ uint32_t compareRealRanks(uint32_t *arr1, uint32_t *arr2, uint32_t arr1_size, ui
 uint32_t cmpGraphAlgorithmsTestStats(void *ref_stats, void *cmp_stats, uint32_t algorithm);
 uint32_t compareDistanceArrays(uint32_t *arr1, uint32_t *arr2, uint32_t arr1_size, uint32_t arr2_size);
 void *runGraphAlgorithmsTest(void *graph, struct Arguments *arguments);
-uint32_t equalFloat(float a, float b, float epsilon);
+uint32_t equalFloat(float a, float b, double epsilon);
 uint32_t compareFloatArrays(float *arr1, float *arr2, uint32_t arr1_size, uint32_t arr2_size);
 
 #endif
